Refactor Send.js
